+++ app/src/main/kotlin/io/treehouses/remote/InitialActivity.kt
@@ -26,6 +26,7 @@ import io.treehouses.remote.callback.HomeInteractListener
 import io.treehouses.remote.callback.NotificationCallback
 import io.treehouses.remote.databinding.ActivityInitial2Binding
 import io.treehouses.remote.ui.home.HomeFragment
+import io.treehouses.remote.ui.services.ServicesFragment
 import io.treehouses.remote.utils.GPSService
 import io.treehouses.remote.utils.LogUtils
