+++ app/src/main/kotlin/io/treehouses/remote/ui/services/ServicesViewModel.kt
@@ -1,0 +1,7 @@
+package io.treehouses.remote.ui.services
+
+import androidx.lifecycle.ViewModel
+
+class ServicesViewModel : ViewModel() {
+
+}
